add form for hacker information
