+++ app/settings.py
@@ -46,6 +46,7 @@ INSTALLED_APPS = [
     'django.contrib.humanize',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'bootstrap3',
     'register',
     'checkin',
     'table'
