+++ app/templates/base.html
@@ -71,9 +71,19 @@
     </div>
 </nav>
 <div class="content container">
-    {% block content %}
-    {% endblock %}
     {% block body %}
+        <div class="margin-top">
+            <div class="col-md-6 col-md-offset-3">
+                <div class="panel panel-default">
+                    <div class="panel-body">
+                        {% block content %}
+                        {% endblock %}
+
+                    </div>
+                </div>
+            </div>
+        </div>
+
     {% endblock %}
 </div>
 </body>
