+++ register/management/commands/insert_missing.py
@@ -1,10 +1,11 @@
 from django.core.management.base import BaseCommand
-from register import forms
+
+from register import typeform
 
 
 class Command(BaseCommand):
     help = 'Fetches and updates forms from Typeform'
 
     def handle(self, *args, **options):
-        fetched = forms.ApplicationsTypeformAll().insert_forms()
+        fetched = typeform.ApplicationsTypeformAll().insert_forms()
         self.stdout.write(self.style.SUCCESS('Successfully fetched %s forms' % len(fetched)))
