+++ register/templates/phases/hacker_info.html
@@ -1,56 +1,3 @@
-<p>Let us know a bit more about yourself</p>
-<form class="form-horizontal">
-    <div class="form-group">
-        <label for="name" class="col-sm-2 control-label">Name</label>
-        <div class="col-sm-10">
-            <input type="email" class="form-control" id="name" placeholder="Bee" required>
-        </div>
-    </div>
-    <div class="form-group">
-        <label for="lastname" class="col-sm-2 control-label">Last name</label>
-        <div class="col-sm-10">
-            <input type="text" class="form-control" id="lastname" placeholder="Biene" required>
-        </div>
-    </div>
-    <div class="form-group">
-        <label for="country" class="col-sm-2 control-label">Nationality</label>
-        <div class="col-sm-10">
-            <input type="text" class="form-control" id="country" placeholder="BeeLand" required>
-        </div>
-    </div>
-    <div class="form-group">
-        <label for="gender" class="col-sm-2 control-label">Gender</label>
-        <div class="col-sm-10">
-            <select class="form-control" id="gender">
-                <option>Male</option>
-                <option>Female</option>
-                <option>Non-binary</option>
-            </select>
-        </div>
-    </div>
-    {#    TODO: Make this an autocomplete with json of all unis#}
-    <div class="form-group">
-        <label for="university" class="col-sm-2 control-label">University/College</label>
-        <div class="col-sm-10">
-            <input type="text" class="form-control" id="university" placeholder="BeeUni" required>
-        </div>
-    </div>
-    <div class="form-group">
-        <label for="graduation_year" class="col-sm-2 control-label">Graduation year</label>
-        <div class="col-sm-10">
-            <input type="number" class="form-control" id="graduation_year" value="2017" step="1" min="2015" required>
-        </div>
-    </div>
-    {#    TODO: Make this an autocomplete with json of all unis#}
-    <div class="form-group">
-        <label for="degree" class="col-sm-2 control-label">Degree pursuing</label>
-        <div class="col-sm-10">
-            <input type="text" class="form-control" id="degree" placeholder="Bee studies" required>
-        </div>
-    </div>
-    <div class="form-group">
-        <div class="col-sm-offset-2 col-sm-10">
-            <button type="submit" class="btn btn-default">Submit</button>
-        </div>
-    </div>
-</form>
+<p>You will only need to fill this once, we will keep this information between editions</p>
+
+{% include 'include/hacker_form.html' %}
