+++ requirements.txt
@@ -1,6 +1,7 @@
 appdirs==1.4.0
 Django==1.10.4
 django-allauth==0.32.0
+django-bootstrap3==8.2.3
 django-datatable==0.2.1
 django-filter==1.0.1
 django-jet==1.0.4
